back-end/requirements.txt
asgiref==3.7.2
avro==1.10.0
avro-python3==1.10.2
boto3==1.26.155
botocore==1.29.165
certifi==2023.7.22
charset-normalizer==3.2.0
confluent-kafka==1.9.0
Django==3.2.4
django-cleanup==8.0.0
django-cors-headers==3.7.0
django-minio-storage==0.5.6
django-storages==1.13.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.1.0
fastavro==1.8.4
idna==3.4
jmespath==1.0.1
minio==7.1.15
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.6.2
six==1.15.0
sqlparse==0.2.4
typing_extensions==4.8.0
urllib3==1.26.16